Add Test

+++ Java/src/main/java/com/guildedrose/balance/BalanceInteractor.java
@@ -10,7 +10,7 @@ public class BalanceInteractor implements BalanceViewer{
     }
 
     @Override
-    public int getBalance() {
+    public double getBalance() {
         return balanceRepository.GetBalance();
     }
 }
